pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ayeye"
version = "0.1.0"
description = "A quick video issue reporter using Gemini"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "google-genai>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"

[project.scripts]
ayeye = "ayeye.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
